% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hela functions.R
\name{b_profile_one}
\alias{b_profile_one}
\title{Compute the profile log-likelihood of the intercept coefficients.}
\usage{
b_profile_one(hypers, train_points, l_train, prior_mean_str)
}
\arguments{
\item{hypers}{A vector containing the hyperparameters of the Gaussian porcess: signal variance and length scale parameters.}

\item{train_points}{A vector containing the training points.}

\item{l_train}{A vector containing the function approximation at the training points.}

\item{prior_mean_str}{A function giving the structure of the prior mean.}
}
\value{
A vector \code{b_prof_log_lik_one} containing the profile log-likelihood
at the given values of hyperparameters and training points.
}
\description{
Compute the profile log-likelihood of the intercept coefficients.
}
