% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hela functions.R
\name{beta_estim_one}
\alias{beta_estim_one}
\title{Compute the estimated intercept parameters}
\usage{
beta_estim_one(train_points, l_train, hypers, p, prior_mean_str)
}
\arguments{
\item{train_points}{A vector containing the training points.}

\item{l_train}{A vector containing the function approximation at the training points.}

\item{hypers}{A vector containing the hyperparameters of the Gaussian porcess: signal variance and length scale parameters.}

\item{p}{The nugget.}

\item{prior_mean_str}{A function giving the structure of the prior mean.}
}
\value{
A list \code{beta_estimates} containing the estimated intercept parameters.
}
\description{
Compute the estimated intercept parameters
of the low-level and high-level.
}
