% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hela functions.R
\name{beta_estim_three}
\alias{beta_estim_three}
\title{Compute the estimated intercept parameters}
\usage{
beta_estim_three(
  train_points_low,
  train_points_mid,
  train_points_high,
  l_train_low,
  l_train_mid,
  l_train_high,
  hypers,
  prior_mean_str
)
}
\arguments{
\item{train_points_low}{A vector containing the training points of the low-level.}

\item{train_points_mid}{A vector containing the training points of the middle-level.}

\item{train_points_high}{A vector containing the training points of the high-level.}

\item{l_train_low}{A vector containing the low-level function approximation at the low-level training points.}

\item{l_train_mid}{A vector containing the middle-level function approximation at the middle-level training points.}

\item{l_train_high}{A vector containing the high-level function approximation at the high-level training points.}

\item{hypers}{A vector containing the hyperparameters of the Gaussian process of each level:
signal variance, length scale and autoregressive parameters.}

\item{prior_mean_str}{A function giving the structure of the prior mean.}
}
\value{
A list \code{beta_estimates} containing the estimated intercept parameters
for the low, middle and high levels.
}
\description{
Compute the estimated intercept parameters
of the low-level, middle-level and high-level.
}
