% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hela functions.R
\name{const_prior}
\alias{const_prior}
\title{Constant prior mean structure of the Gaussian process}
\usage{
const_prior(input_points)
}
\arguments{
\item{input_points}{A vector containing the input points.}
}
\value{
A matrix containing the constant prior mean structure used for the Gaussian process regression.
}
\description{
Constant prior mean structure of the Gaussian process
}
