% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hela functions.R
\name{cov_fun_three}
\alias{cov_fun_three}
\title{Compute the covariance matrix of the Gaussian process posterior.}
\usage{
cov_fun_three(hypers, train_points_low, train_points_mid, train_points_high)
}
\arguments{
\item{hypers}{A vector containing the hyperparameters of the Gaussian process of each level:
signal variance, length scale and autoregressive parameters.}

\item{train_points_low}{A vector containing the training points of the low-level.}

\item{train_points_mid}{A vector containing the training points of the middle-level.}

\item{train_points_high}{A vector containing the training points of the high-level.}
}
\value{
A block matrix \code{cov_matr_block} containing the posterior covariance of the
GP posterior for the three-level approximation.
}
\description{
Compute the ovariance matrix of the three-level Gaussian process posterior.
}
