% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hela functions.R
\name{cov_V}
\alias{cov_V}
\title{Compute the block covariance matrix V.}
\usage{
cov_V(train_points_low, train_points_high, hypers)
}
\arguments{
\item{train_points_low}{A vector containing the training points of the low-level.}

\item{train_points_high}{A vector containing the training points of the high-level.}

\item{hypers}{A vector containing the hyperparameters of the Gaussian porcess of each level:
signal variance, length scale and autoregressive parameters.}
}
\value{
A block matrix \code{V} containing the kernel matrices for both levels of
approximation.
}
\description{
Compute the block covariance matrix V of the two level Gaussian process.
}
