% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hela functions.R
\name{egu_three}
\alias{egu_three}
\title{Expected gain in utility for thee-level Gaussian processes}
\usage{
egu_three(
  result_gp,
  hypers,
  hyper_initial,
  egu_min,
  max_points,
  cand_points,
  log_lik_high,
  log_lik_mid,
  log_lik_low,
  train_points_low,
  train_points_mid,
  train_points_high,
  l_train_low,
  l_train_mid,
  l_train_high,
  test_points,
  cost_h,
  cost_m,
  cost_l,
  prior_mean_str
)
}
\arguments{
\item{result_gp}{A list containing the posterior mean vector and the posterior covariance matrix of the
multi-level GP posterior distribution.}

\item{hypers}{A vector containing the initial values of the hyperparameters,
used for optimisation, of the Gaussian process of each level:
signal variance, length scale and autoregressive parameters.}

\item{hyper_initial}{A vector containing the initial values of the hyperparameters for estimation.}

\item{egu_min}{Expected gain in utility minimum tolerance as a stopping rule.}

\item{max_points}{Maximum points to be added to the experimental design as a stopping rule.}

\item{cand_points}{A vector containing the candidate points.}

\item{log_lik_high}{Function of the high-level points giving the high-level approximation.}

\item{log_lik_mid}{Function of the middle-level points giving the middle-level approximation.}

\item{log_lik_low}{Function of the low-level points giving the low-level approximation.}

\item{train_points_low}{A vector containing the training points of the low-level.}

\item{train_points_mid}{A vector containing the training points of the middle-level.}

\item{train_points_high}{A vector containing the training points of the high-level.}

\item{l_train_low}{A vector containing the low-level function approximation at the low-level training points.}

\item{l_train_mid}{A vector containing the middle-level function approximation at the middle-level training points.}

\item{l_train_high}{A vector containing the high-level function approximation at the high-level training points.}

\item{test_points}{A vector containing the test points.}

\item{cost_h}{Constant indicating the cost of the high-level.}

\item{cost_m}{Constant indicating the cost of the middle-level.}

\item{cost_l}{Constant indicating the cost of the low-level.}

\item{prior_mean_str}{A function giving the structure of the prior mean.}
}
\value{
A list containing the estimated hyperparameters of each level, \code{ssquare1}, \code{ssquare2}, \code{ssquare3},
\code{l1}, \code{l2}, \code{l3}, the autoregressive parameters \code{r1} and \code{r2},
the training points of each level, \code{train_points_low}, \code{train_points_mid} and \code{train_points_high},
and the function evaluations \code{l_train_low}, \code{l_train_mid} and \code{l_train_high},
the EGU per cost for each level \code{egu_low}, \code{egu_mid} and \code{egu_high}
after the addition of new points and the new points added to the design, \code{new_point}.
}
\description{
Compute the experimental design of each level of approximation using expected gain in utility.
}
