% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hela functions.R
\name{gp_post}
\alias{gp_post}
\title{Compute the multi-level Gaussian process posterior.}
\usage{
gp_post(
  train_points_low,
  train_points_high,
  l_train_low,
  l_train_high,
  test_points,
  hypers,
  prior_mean_str,
  approx_levels,
  train_points_mid = NULL,
  l_train_mid = NULL
)
}
\arguments{
\item{train_points_low}{A vector containing the training points of the low-level.}

\item{train_points_high}{A vector containing the training points of the high-level.}

\item{l_train_low}{A vector containing the low-level function approximation at the low-level training points.}

\item{l_train_high}{A vector containing the high-level function approximation at the high-level training points.}

\item{test_points}{A vector containing the test points used for prediction.}

\item{hypers}{A vector containing the hyperparameters of the Gaussian process of each level:
signal variance, length scale and autoregressive parameters.}

\item{prior_mean_str}{A function giving the structure of the prior mean.}

\item{approx_levels}{Constant indicating how many levels of approximation to use. Choose between 2 or 3.}

\item{train_points_mid}{A vector containing the training points of the middle-level for a three-level approximation.
Default value is \code{NULL}.}

\item{l_train_mid}{A vector containing the middle-level function approximation at the middle-level training points.
for a three-level approximation. Default value is \code{NULL}.}
}
\value{
A list containing the posterior mean vector \code{gp_mean} and
the posterior covariance matrix \code{gp_cov}.
}
\description{
Compute the mean vector and the covariance matrix of the two level Gaussian process posterior.
}
