% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hela functions.R
\name{gp_posterior}
\alias{gp_posterior}
\title{Compute the single level Gaussian process posterior.}
\usage{
gp_posterior(
  train_points,
  test_points,
  sigma_sq,
  lambda,
  l_train,
  p,
  prior_mean_str = NULL
)
}
\arguments{
\item{train_points}{A vector containing the training points.}

\item{test_points}{A vector containing the test points.}

\item{sigma_sq}{Signal variance parameter of the Gaussian process.}

\item{lambda}{Length scale parameter of the Gaussian process.}

\item{l_train}{A vector containing the function approximation at the training points.}

\item{p}{The nugget.}

\item{prior_mean_str}{A function giving the structure of the prior mean.}
}
\value{
A list \code{posterior} containing the posterior mean vector \code{mean} and
the posterior covariance matrix \code{covariance}.
}
\description{
Compute the mean vector and the covariance matrix of the two level Gaussian process posterior.
}
