% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hela functions.R
\name{gp_posterior_plot}
\alias{gp_posterior_plot}
\title{Gaussian process posterior plot for the two-level approximation}
\usage{
gp_posterior_plot(
  train_points_low,
  train_points_high,
  test_points,
  l_train_low,
  l_train_high,
  gp_mean,
  gp_cov,
  accurate = NULL,
  train_points_mid = NULL,
  l_train_mid = NULL
)
}
\arguments{
\item{train_points_low}{A vector containing the training points of the low-level.}

\item{train_points_high}{A vector containing the training points of the high-level.}

\item{test_points}{A vector containing the test points used for prediction.}

\item{l_train_low}{A vector containing the low-level function approximation at the low-level training points.}

\item{l_train_high}{A vector containing the high-level function approximation at the high-level training points.}

\item{gp_mean}{A vector containing Gaussian process posterior mean - result of the \code{gp_post()} function.}

\item{gp_cov}{A matrix containing Gaussian process posterior covariance - result of the \code{gp_post()} function.}

\item{accurate}{Exact - Accurate approximation of the high-level if available. Default value is \code{NULL}}

\item{train_points_mid}{A vector containing the training points of the middle-level.
Default value is \code{NULL}. Used when using three levels of approximation.}

\item{l_train_mid}{A vector containing the middle-level function approximation at the middle-level training points.
Default value is \code{NULL} Used when using three levels of approximation.}
}
\value{
Plot of the GP posterior distribution.
}
\description{
Plot the Gaussian process posterior mean and credible intervals for the two-level approximation.
}
