% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hela functions.R
\name{linear_prior}
\alias{linear_prior}
\title{Linear prior mean structure of the Gaussian process}
\usage{
linear_prior(input_points)
}
\arguments{
\item{input_points}{A vector containing the input points.}
}
\value{
A matrix containing the linear prior mean structure used for the Gaussian process regression.
}
\description{
Linear prior mean structure of the Gaussian process
}
