% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hela functions.R
\name{mean_fun_two}
\alias{mean_fun_two}
\title{Compute the mean vector of the Gaussian process posterior.}
\usage{
mean_fun_two(
  hypers,
  train_points_low,
  train_points_high,
  l_train_low,
  l_train_high,
  prior_mean_str
)
}
\arguments{
\item{hypers}{A vector containing the hyperparameters of the Gaussian process of each level:
signal variance, length scale and autoregressive parameters.}

\item{train_points_low}{A vector containing the training points of the low-level.}

\item{train_points_high}{A vector containing the training points of the high-level.}

\item{l_train_low}{A vector containing the low-level function approximation at the low-level training points.}

\item{l_train_high}{A vector containing the high-level function approximation at the high-level training points.}

\item{prior_mean_str}{A function giving the structure of the prior mean.}
}
\value{
A vector \code{mean_vec} containing the posterior mean vector of the
GP posterior for the two-level approximation.
}
\description{
Compute the mean vector of the two level Gaussian process posterior.
}
