% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hela functions.R
\name{quad_prior}
\alias{quad_prior}
\title{Quadratic prior mean structure of the Gaussian process}
\usage{
quad_prior(input_points)
}
\arguments{
\item{input_points}{A vector containing the input points.}
}
\value{
A matrix containing the quadratic prior mean structure used for the Gaussian process regression.
}
\description{
Quadratic prior mean structure of the Gaussian process
}
